Added a very early COMPASS SYSTEM to the game. WIP.
A rough simple compass system (still in development) with working 360 degree accurate bearings of N-E-S-W.

The debug_HUD Aidan made had to be removed from the EventBeginPlay graph as it would not allow the compass to my knowledge show up with it.

I'm sure you (Aidan) can reapply it if need be with my new compass sytem added.

CT.